backend/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7
python-engineio==4.8.0
python-socketio==5.10.0
eventlet==0.33.3
openai>=1.12.0
